+++ CKE/config.js
@@ -1,5 +1,8 @@
 CKEDITOR.editorConfig = function (config) {
 
+    // clear cache?
+    //CKEDITOR.timestamp='ABCD';
+
     config.autoGrow_onStartup = true;
     config.autoGrow_bottomSpace = 20;